create lib directory, move getFeeds into it
